% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_vnp_avg_help.R
\name{getCov}
\alias{getCov}
\title{Get empirical covariance matrix from a updateCov() object}
\usage{
getCov(covObj, eps = 1e-06)
}
\description{
Get empirical covariance matrix from a updateCov() object
}
\keyword{internal}
