% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_vnpc_avg.R
\name{gibbs_vnpc_avg}
\alias{gibbs_vnpc_avg}
\title{Gibbls sampler for VNP based on average periodogram}
\usage{
gibbs_vnpc_avg(
  mpg_avg,
  f_param_avg,
  Nb,
  Ntotal,
  burnin,
  thin = 1,
  print_interval = 100,
  numerical_thresh = 1e-07,
  adaption.N = burnin,
  adaption.batchSize = 50,
  adaption.tar = 0.44,
  eta = dim(mpg_avg)[2],
  omega = dim(mpg_avg)[2],
  Sigma = 10000 * diag(dim(mpg_avg)[2]),
  k.theta = 0.01,
  kmax = 100 * coars + 500 * (!coars),
  trunc_l = 0.1,
  trunc_r = 0.9,
  coars = F,
  L = max(20, (dim(mpg_avg)[1] * 2)^(1/3)),
  sqrt_d = F
)
}
\description{
Gibbs sampler for multivariate Bayesian nonparametric inference with corrected Whittle likelihood
}
