% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{llike_corrected_avg}
\alias{llike_corrected_avg}
\title{Log likelihood of a pseudo parametric VAR model}
\usage{
llike_corrected_avg(mpg_avg, f, f_param_avg_half, Nb)
}
\description{
Log likelihood of a pseudo parametric VAR model
}
\keyword{internal}
