% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_gamma_matrix_subordinator_core.R
\name{llike_matrixGamma}
\alias{llike_matrixGamma}
\title{Likelihood computation}
\usage{
llike_matrixGamma(
  samp_freq,
  omega,
  mpg_avg,
  Nb,
  FZ,
  r,
  U,
  Z,
  k,
  db.list,
  bspline,
  degree_bs,
  V,
  W,
  corrected,
  f_param_avg_half,
  phi,
  sigma_ar,
  prior.q,
  prior.cholesky,
  excludeBoundary,
  verbose
)
}
\description{
Likelihood computation
}
\keyword{internal}
