% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_multivariate_util.R
\name{llike_var_corrected}
\alias{llike_var_corrected}
\title{The corrected likelihood}
\usage{
llike_var_corrected(FZ, ar, f_param_half, sigma, f, excludeBoundary = T)
}
\description{
The corrected likelihood
}
\keyword{internal}
