% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_multivariate_util.R
\name{llike_var_corrected_q}
\alias{llike_var_corrected_q}
\title{The corrected likelihood using the q-parametrisation}
\usage{
llike_var_corrected_q(
  FZ,
  ar,
  f_param_half,
  f_param_half_trans,
  sigma,
  q,
  sqrt_d,
  excludeBoundary = T
)
}
\description{
The corrected likelihood using the q-parametrisation
}
\keyword{internal}
