% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{llike_whittle}
\alias{llike_whittle}
\title{log Whittle likelihood (multivariate), unnormalized
Note: omega=0 and omega=pi excluded!}
\usage{
llike_whittle(FZ, f)
}
\description{
log Whittle likelihood (multivariate), unnormalized
Note: omega=0 and omega=pi excluded!
}
\keyword{internal}
