% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{llike_whittle_avg}
\alias{llike_whittle_avg}
\title{The log Whittle likelihood for the averaged periodogram}
\usage{
llike_whittle_avg(mpg_avg, f, Nb)
}
\description{
The log Whittle likelihood for the averaged periodogram
}
\keyword{internal}
