% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{llike_whittle_sum}
\alias{llike_whittle_sum}
\title{Sum of the segmented log Whittle likelihood over all frequencies
Based on the complex Wishart distribution}
\usage{
llike_whittle_sum(FZ, f)
}
\description{
Sum of the segmented log Whittle likelihood over all frequencies
Based on the complex Wishart distribution
}
\keyword{internal}
