% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logdet_cube_sum}
\alias{logdet_cube_sum}
\title{Sum of log determinants of a matrix array excluding boundaries}
\usage{
logdet_cube_sum(f)
}
\description{
Sum of log determinants of a matrix array excluding boundaries
}
\keyword{internal}
