% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_vnp_avg_help.R
\name{updateCov}
\alias{updateCov}
\title{Function for updating the proposal covariance}
\usage{
updateCov(X, covObj = NA)
}
\description{
Function for updating the proposal covariance
}
\keyword{internal}
