% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{varma_transfer2psd_cpp}
\alias{varma_transfer2psd_cpp}
\title{Get VARMA PSD from transfer polynomials 
Helping function for \code{psd_varma_cpp}}
\usage{
varma_transfer2psd_cpp(transfer_ar, transfer_ma, sigma)
}
\description{
Get VARMA PSD from transfer polynomials 
Helping function for \code{psd_varma_cpp}
}
\keyword{internal}
