% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_gamma_matrix_subordinator.R
\name{gibbs_m_avg_nuisance}
\alias{gibbs_m_avg_nuisance}
\title{Title}
\usage{
gibbs_m_avg_nuisance(
  data,
  mcmc_params,
  seg_n = 1,
  truncation = FALSE,
  trunc_freq = NULL,
  trunc_freq_lim = NULL,
  corrected,
  prior_params,
  model_params
)
}
\description{
Title
}
\keyword{internal}
