% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_vnpc_avg.R
\name{gibbs_vnpc_avg}
\alias{gibbs_vnpc_avg}
\title{Gibbls sampler for VNP based on average periodogram}
\usage{
gibbs_vnpc_avg(
  mpg_avg,
  f_param_avg,
  Nb,
  Ntotal,
  burnin,
  thin = 1,
  print_interval = 100,
  numerical_thresh = 1e-07,
  adaption.N = burnin,
  adaption.batchSize = 50,
  adaption.tar = 0.44,
  eta = dim(mpg_avg)[2],
  omega = dim(mpg_avg)[2],
  Sigma = 10000 * diag(dim(mpg_avg)[2]),
  k.theta = 0.01,
  kmax = 100 * coars + 500 * (!coars),
  trunc_l = 0.1,
  trunc_r = 0.9,
  coars = F,
  L = max(20, (dim(mpg_avg)[1] * 2)^(1/3)),
  sqrt_d = F
)
}
\arguments{
\item{mpg_avg}{Averaged periodogram.}

\item{f_param_avg}{Averaged PSD from the prefitted parametric model.}

\item{Nb}{Number of blocks.}

\item{Ntotal}{Total number of MCMC iterations.}

\item{burnin}{Number of initial MCMC iterations to discard.}

\item{thin}{Thinning interval for saved samples.}

\item{print_interval}{Interval for printing progress updates.}

\item{numerical_thresh}{Numerical threshold used for stability checks.}

\item{adaption.N}{Number of iterations used during adaptive tuning.}

\item{adaption.batchSize}{Batch size used during adaptive tuning.}

\item{adaption.tar}{Target acceptance rate used during adaptive tuning.}

\item{eta}{AGamma process parameter, real number > ncol(data) - 1.}

\item{omega}{AGamma process parameter, positive constant.}

\item{Sigma}{AGamma process parameter, Hpd matrix.}

\item{k.theta}{Prior parameter for polynomial degree k (propto exp(-k.theta*k*log(k))).}

\item{kmax}{Upper bound for polynomial degree of Bernstein-Dirichlet mixture (can be set to Inf, algorithm is faster with kmax<Inf due to pre-computation of basis functions, but values 500<kmax<Inf are very memory intensive).}

\item{trunc_l, trunc_r}{Left and right truncation of Bernstein polynomial basis functions, 0<=trunc_l<trunc_r<=1.}

\item{coars}{Flag indicating whether coarsened or default Bernstein polynomials are used (see Appendix E.1 in Ghosal and van der Vaart 2017).}

\item{L}{Truncation parameter of Gamma process.}

\item{sqrt_d}{Flag indicating whether using the regular squre root of a positive definite matrix. Otherwise, using the Cholesky decomposition.}
}
\value{
list containing the following fields:
}
\description{
Gibbs sampler for multivariate Bayesian nonparametric inference with corrected Whittle likelihood
}
